+++ benchmarking/async-prepared-select-ruby-driver.rb
@@ -22,7 +22,7 @@ require 'cassandra'
 class PreparedSelectRubyDriver < Benchmark
     def connect_to_cluster
         puts "#{Time.now - start} Connecting to cluster..."
-        @cluster = Cassandra.connect(hosts: ['127.0.0.1'])
+        @cluster = Cassandra.cluster(hosts: ['127.0.0.1'])
         @session = @cluster.connect("simplex")
         @statement  = @session.prepare('SELECT * FROM songs')
     end