+++ integration/load_balancing/round_robin_test.rb
@@ -34,7 +34,7 @@ class RoundRobinTest < IntegrationTestCase
   def test_round_robin_used_explicitly
     setup_schema
     policy  = Cassandra::LoadBalancing::Policies::RoundRobin.new
-    cluster = Cassandra.connect(load_balancing_policy: policy)
+    cluster = Cassandra.cluster(load_balancing_policy: policy)
     session = cluster.connect("simplex")
 
     hosts_used = []
@@ -52,7 +52,7 @@ class RoundRobinTest < IntegrationTestCase
     allowed_ips = ["127.0.0.1", "127.0.0.3"]
     round_robin = Cassandra::LoadBalancing::Policies::RoundRobin.new
     whitelist = Cassandra::LoadBalancing::Policies::WhiteList.new(allowed_ips, round_robin)
-    cluster = Cassandra.connect(load_balancing_policy: whitelist)
+    cluster = Cassandra.cluster(load_balancing_policy: whitelist)
     session = cluster.connect("simplex")
 
     hosts_used = []
@@ -68,7 +68,7 @@ class RoundRobinTest < IntegrationTestCase
   def test_round_robin_ignores_datacenters
     setup_schema
     policy = Cassandra::LoadBalancing::Policies::RoundRobin.new
-    cluster = Cassandra.connect(load_balancing_policy: policy)
+    cluster = Cassandra.cluster(load_balancing_policy: policy)
     session = cluster.connect("simplex")
 
     hosts_used = []
@@ -85,7 +85,7 @@ class RoundRobinTest < IntegrationTestCase
     setup_schema
     datacenter = "dc1"
     policy = Cassandra::LoadBalancing::Policies::DCAwareRoundRobin.new(datacenter)
-    cluster = Cassandra.connect(load_balancing_policy: policy)
+    cluster = Cassandra.cluster(load_balancing_policy: policy)
     session = cluster.connect("simplex")
 
     hosts_used = []
@@ -102,7 +102,7 @@ class RoundRobinTest < IntegrationTestCase
     setup_schema
     datacenter = "dc2"
     policy = Cassandra::LoadBalancing::Policies::DCAwareRoundRobin.new(datacenter)
-    cluster = Cassandra.connect(consistency: :one, load_balancing_policy: policy)
+    cluster = Cassandra.cluster(consistency: :one, load_balancing_policy: policy)
     session = cluster.connect("simplex")
 
     @@ccm_cluster.stop_node('node3')
@@ -122,7 +122,7 @@ class RoundRobinTest < IntegrationTestCase
     setup_schema
     datacenter = "dc1"
     policy = Cassandra::LoadBalancing::Policies::DCAwareRoundRobin.new(datacenter)
-    cluster = Cassandra.connect(consistency: :one, load_balancing_policy: policy)
+    cluster = Cassandra.cluster(consistency: :one, load_balancing_policy: policy)
     session = cluster.connect("simplex")
 
     @@ccm_cluster.stop_node('node1')
@@ -141,7 +141,7 @@ class RoundRobinTest < IntegrationTestCase
     datacenter = "dc2"
     remotes_to_try = 1
     policy = Cassandra::LoadBalancing::Policies::DCAwareRoundRobin.new(datacenter, remotes_to_try)
-    cluster = Cassandra.connect(consistency: :one, load_balancing_policy: policy)
+    cluster = Cassandra.cluster(consistency: :one, load_balancing_policy: policy)
     session = cluster.connect("simplex")
 
     @@ccm_cluster.stop_node('node3')
@@ -163,7 +163,7 @@ class RoundRobinTest < IntegrationTestCase
     setup_schema
     datacenter = "dc2"
     policy = Cassandra::LoadBalancing::Policies::DCAwareRoundRobin.new(datacenter)
-    cluster = Cassandra.connect(consistency: :one, load_balancing_policy: policy)
+    cluster = Cassandra.cluster(consistency: :one, load_balancing_policy: policy)
     session = cluster.connect("simplex")
 
     @@ccm_cluster.stop_node('node3')
@@ -181,7 +181,7 @@ class RoundRobinTest < IntegrationTestCase
     datacenter = "dc2"
     use_remote = true
     policy = Cassandra::LoadBalancing::Policies::DCAwareRoundRobin.new(datacenter, nil, use_remote)
-    cluster = Cassandra.connect(consistency: :one, load_balancing_policy: policy)
+    cluster = Cassandra.cluster(consistency: :one, load_balancing_policy: policy)
     session = cluster.connect("simplex")
 
     @@ccm_cluster.stop_node('node3')