+++ integration/security/ssl_encryption_test.rb
@@ -30,7 +30,7 @@ class SSLEncryptionTest < IntegrationTestCase
   end
 
   def test_can_connect_with_default_ssl
-    cluster = Cassandra.connect(ssl: true)
+    cluster = Cassandra.cluster(ssl: true)
     refute_nil cluster
   ensure
     cluster.close
@@ -38,13 +38,13 @@ class SSLEncryptionTest < IntegrationTestCase
 
   def test_raise_error_when_not_using_ssl
     assert_raises(Cassandra::Errors::NoHostsAvailable) do
-      cluster = Cassandra.connect
+      cluster = Cassandra.cluster
       cluster.close
     end
   end
 
   def test_can_connect_with_ssl_ca
-    cluster = Cassandra.connect(server_cert: @@server_cert)
+    cluster = Cassandra.cluster(server_cert: @@server_cert)
     refute_nil cluster
   ensure
     cluster.close
@@ -52,7 +52,7 @@ class SSLEncryptionTest < IntegrationTestCase
 
   def test_raise_error_on_invalid_ca_provided
     assert_raises(Cassandra::Errors::NoHostsAvailable) do
-      cluster = Cassandra.connect(server_cert: '')
+      cluster = Cassandra.cluster(server_cert: '')
       cluster.close
     end
   end