+++ lib/cassandra/session.rb
@@ -54,7 +54,7 @@ module Cassandra
     #   relevant for conditional updates and specifies a serial consistency to
     #   be used, one of {Cassandra::SERIAL_CONSISTENCIES}
     #
-    # @see Cassandra.connect Options that can be specified on the cluster-level
+    # @see Cassandra.cluster Options that can be specified on the cluster-level
     #   and their default values.
     #
     # @note Last argument will be treated as `options` if it is a {Hash}.